Refactor AddToCalendar file architecture
